+++ bridge/Stats.h
@@ -49,8 +49,8 @@ struct MixerManagerStats
     uint32_t _receivePoolSize = 0;
     uint32_t _sendPoolSize = 0;
     uint32_t _udpSharedEndpointsSendQueue = 0;
-    double _udpSharedEndpointsReceiveKbps = 0.0;
-    double _udpSharedEndpointsSendKbps = 0.0;
+    uint32_t _udpSharedEndpointsReceiveKbps = 0;
+    uint32_t _udpSharedEndpointsSendKbps = 0;
 
     MixerManagerStats() {}
